2025.06.06 / 관리자 페이지 및 각종 페이지 기능 완성 및 수정, 따로 오류 문제는 관리자 페이지에서 상태 변경 오류 및 글 작성 후 올라가지 않는 버그 있음. 따로 수정 예정.

+++ src/main/java/com/Ignis/admin/AdminRestController.java
@@ -1,13 +1,20 @@
 package com.Ignis.admin;
 
-import com.Ignis.common.enums.Status;
-import com.Ignis.donation.bo.DonationBO;
-import com.Ignis.post.bo.PostBO;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
 import java.util.HashMap;
 import java.util.Map;
+
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.Ignis.common.enums.Status;
+import com.Ignis.home.donation.bo.DonationBO;
+import com.Ignis.post.bo.PostBO;
+
+import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequestMapping("/admin")
@@ -28,7 +35,7 @@ public class AdminRestController {
     }
 
     @PostMapping("/donation-status-update")
-    public String updateDonationStatus(@RequestParam("donationId") int donationId,
+    public String updateDonationStatus(@RequestParam("donationId") Long donationId,
                                        @RequestParam("status") Status status) {
         donationBO.updateDonationStatus(donationId, status);
         return "상태 변경 완료";
